% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{calc_km}
\alias{calc_km}
\title{Kaplan-Meier curve estimation}
\usage{
calc_km(df, assume_all_died = FALSE)
}
\description{
Kaplan-Meier curve estimation
}
\details{
df should have columns tail, n_event, n_died (n_event includes n_died).

If assume_all_died=TRUE, we set n_died = n_event.
}
