% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{count_umis}
\alias{count_umis}
\title{Count UMIs (or reads if UMIs not present)}
\usage{
count_umis(sited_reads)
}
\description{
Reads are counted even if they don't have a tail. Use for examining site expression levels.
}
