% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_fastqs.R
\name{demux_reads}
\alias{demux_reads}
\title{Demultiplex into FASTQ files from parquet file}
\usage{
demux_reads(out_dir, in_file, sample_names = NULL)
}
\description{
Demultiplex into FASTQ files from parquet file
}
