% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ingest_tt}
\alias{ingest_tt}
\title{Ingest Tail Tools output}
\usage{
ingest_tt(
  out_dir,
  in_dir,
  tail_source,
  read_pairs_file = NULL,
  site_pad = 10,
  min_tail = 19,
  length_trim = 10,
  limit = NA,
  steps = 1:6
)
}
\arguments{
\item{tail_source}{Should be one of "tt", "read1", or "read2". If "tt", Tail Tools tail lengths are used. If "read1", poly(A) tail lengths from read 1 are used. If "read2", poly(T) tail lengths from read 2 are used. If using "read1" or "read2", read_pairs_file must be given.}
}
\description{
Ingest Tail Tools output
}
