% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{km_quantile_bound}
\alias{km_quantile_bound}
\title{Quantile on some number of standard errors from the Kaplan-Meier curve}
\usage{
km_quantile_bound(km, prop, z)
}
\description{
Quantile on some number of standard errors from the Kaplan-Meier curve
}
