% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_bam_into}
\alias{load_bam_into}
\title{Load data from a BAM file}
\usage{
load_bam_into(
  dest_file,
  bam_file,
  tail_source,
  read_pairs_file = NULL,
  clips_file = NULL,
  limit = NA,
  filter_secondary = TRUE
)
}
\description{
Load data from a BAM file. Information about genomic length and tail length is included. One alignment from each multimapping read is kept. Only the 3' end position is kept.
}
\details{
Multimappers are detected using NH attribute, and there should be only one alignment that isn't flagged as a "secondary alignment". This is what STAR does by default.

Tail Tools BAM files don't require this "secondary alignment" filtering step, so it can and should be disabled in this case.
}
