% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_tt_sample_into}
\alias{load_tt_sample_into}
\title{Load data from a Tail Tools sample directory}
\usage{
load_tt_sample_into(
  dest_filename,
  path,
  tail_source,
  read_pairs_file = NULL,
  limit = NA
)
}
\description{
Load data from a Tail Tools sample directory. Information about genomic length and tail length is included. Multimappers are discarded. Only the 3' end position is kept.
}
