% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_tt_sites}
\alias{load_tt_sites}
\title{Load site information from Tail Tools output}
\usage{
load_tt_sites(path, reference_gff = NULL, reference_fasta = NULL)
}
\value{
A tibble.

pos is the final base of the transcript in the genome (1-based coordinates).
}
\description{
Load site information from Tail Tools output
}
