% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_constructor.R
\name{make_tests_oneway}
\alias{make_tests_oneway}
\title{Make a set of tests with one experimental factor}
\usage{
make_tests_oneway(samples, groups, batches = NULL)
}
\arguments{
\item{samples}{Chacter vector of sample names.}

\item{groups}{Factor specifying experimental groups of interest.}

\item{batches}{An option batch factor.}
}
\description{
Make a set of tests with one experimental factor
}
