% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_constructor.R
\name{make_tests_twoway}
\alias{make_tests_twoway}
\title{Make a set of tests with one experimental factors}
\usage{
make_tests_twoway(samples, groups1, groups2, batches = NULL)
}
\arguments{
\item{samples}{Chacter vector of sample names.}

\item{groups1}{Second factor in the experiment.}

\item{batches}{An option batch factor.}
}
\description{
Make a set of tests with one experimental factors
}
