% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunked.R
\name{map_bam_chunks}
\alias{map_bam_chunks}
\title{Map GAlignments chunks}
\usage{
map_bam_chunks(filename, param, callback, limit = NA, chunk = 5e+06)
}
\description{
Note: If limit is larger than chunk size, it is rounded up to a multiple of chunk size.
}
\details{
TODO: Also provide a scan_ version.
}
