% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_fastqs.R
\name{reads_peek}
\alias{reads_peek}
\title{Examine a random selection of reads}
\usage{
reads_peek(out_file, in_dir, n = 100, line_width = 120, seed = 563)
}
\description{
Examine a random selection of reads
}
