% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_fastqs.R
\name{reads_report}
\alias{reads_report}
\title{Reads report}
\usage{
reads_report(in_file, out_file = paste0(in_file, ".report.html"))
}
\description{
Generate an HTML report for a parquet file produced by \code{ingest_reads()}.
}
