% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_tq}
\alias{run_tq}
\title{Run tailquant pipeline}
\usage{
run_tq(
  out_dir,
  in_dir,
  sample_names,
  reference,
  extension,
  noncoding_prop = 1/3,
  clip_min_untemplated = 12,
  clip_min_length = 20,
  max_t_read_1 = 20,
  min_t = 0,
  min_tail = 13,
  tail_source = "read2",
  length_trim = 0,
  site_upstrand = 300,
  site_pad = 10,
  site_suppression_pad = 50,
  site_min_reads = 50,
  tail_excess_required = 5,
  a_prop = 0.6,
  keep_multimappers = TRUE,
  steps = 1:11
)
}
\arguments{
\item{tail_excess_required}{A site will only be used if the mean tail length exceeds the downstrand genomic A length by this much. This provides a filter for mispriming.}

\item{a_prop}{When finding the genomic A length downstrand of a A, require this level of A purity.}
}
\description{
Run tailquant pipeline
}
