% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunked.R
\name{scan_bam_chunks}
\alias{scan_bam_chunks}
\title{Scan GAlignments chunks}
\usage{
scan_bam_chunks(filename, param, callback, limit = NA, chunk = 1e+06)
}
\description{
Note: If limit is larger than chunk size, it is rounded up to a multiple of chunk size.
}
