% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunked.R
\name{scan_parquet}
\alias{scan_parquet}
\title{Scan a parquet file}
\usage{
scan_parquet(filename, callback, columns = NULL, message = "Processing")
}
\description{
Note it may be necessary for your callback to invoke gc() to avoid excess memory usage.
}
