% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{site_reads}
\alias{site_reads}
\title{Join reads to sites}
\usage{
site_reads(reads, sites, site_pad)
}
\arguments{
\item{site_pad}{Reads ending within \link{position-site_pad, position+site_pad} will be counted to a site.}
}
\description{
Join reads to sites
}
