% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_constructor.R
\name{tests_to_tt}
\alias{tests_to_tt}
\title{Convert tailquant test list to Tail Tools format}
\usage{
tests_to_tt(tests, pipeline_dir)
}
\arguments{
\item{tests}{A named list of test specifications, for example as produced by make_tests_oneway or make_tests_twoway.}

\item{pipeline_dir}{The Tail Tools pipeline output directory.}
}
\description{
Convert a list of test specifications for tailquant to the format used by the older Tail Tools software.
}
