% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{tq_proportions}
\alias{tq_proportions}
\title{Get matrix of proprtion of tails as long or longer than a certain length}
\usage{
tq_proportions(tq, tail)
}
\description{
Get matrix of proprtion of tails as long or longer than a certain length
}
