% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{tq_warmup}
\alias{tq_warmup}
\title{Ensure some computations are cached.}
\usage{
tq_warmup(tq)
}
\arguments{
\item{tq}{A TailQuant object.}
}
\description{
Ensure some computations are cached, specifically those likely to be used by the Shiny app.
}
