% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_fastqs.R
\name{ingest_read_pairs}
\alias{ingest_read_pairs}
\title{Initial loading and processing of Pooled PAT-Seq data into a parquet file}
\usage{
ingest_read_pairs(
  out_prefix,
  reads1,
  reads2,
  samples = NULL,
  max_mismatch = 1,
  clip_quality_char = "I",
  clip_penalty = 4,
  poly_penalty = 1000,
  suffix_penalty = 4,
  limit = Inf,
  chunk_size = 2e+07,
  decompressor = "unpigz",
  compressor = "pigz --fast"
)
}
\arguments{
\item{decompressor}{Decompression program to use (with any needed flags). Input will be piped to this program. Use \code{"cat"} if FASTQ files are not compressed.}

\item{compressor}{Compression program to use after splitting. Input will be piped to this program.}
}
\description{
Extract barcodes, UMIs, clipping information and poly(A)/poly(T) lengths, producing a large parquet file.
}
\details{
By default assumes FASTQ files are gzipped and uses unpigz to decompress and pigz to recompress. pigz must be installed for this to work.
}
