% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_fastqs.R
\name{ingest_read_pairs}
\alias{ingest_read_pairs}
\title{Initial loading and processing of Pooled PAT-Seq data into a parquet file}
\usage{
ingest_read_pairs(
  output_filename,
  reads1,
  reads2,
  samples,
  max_mismatch = 1,
  clip_quality_char = "I",
  clip_penalty = 4,
  poly_penalty = 1000,
  suffix_penalty = 4,
  limit = Inf
)
}
\description{
Initial loading and processing of Pooled PAT-Seq data into a parquet file
}
