% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_exploratory.R
\name{tq_plot_sample_tails}
\alias{tq_plot_sample_tails}
\title{Plot a distribution of tail lengths per sample (or group)}
\usage{
tq_plot_sample_tails(tq, grouping = NULL)
}
\description{
Plot a distribution of tail lengths per sample (or group)
}
